Non-looping DFS over the call-graph.
Extracts paths through the code from entry points to functions.

The CallGraph has not identified main as the entry point so every function with external
linkage is being iterated as a possible entry point. The handling of the entry point as
a null Function* is a bit fishy. Needs a bit of clean-up.