Flow graph for interesting functions
This is a combination of the control- and data- flow graphs for the supplied
function. It is a bi-partite graph, the circled nodes are data-values and the
shape-less nodes are instructions. The solid edges show data-flow and the
dashed-lines are control flow. The tree of ancestors on this graph is a slice
of the program affecting the target node.

It is a bit busy at the moment; control-edges that are redundant because of
existing data-flow edges should be pruned.