Attempt to construct the bi-partite digraph.
Two graphs are produced from the function: data-flow reachability and a pruned
control-flow graph. The df-reachability should be the transitive closure of the
data-flow relation in the function. The control-flow is pruned to remove control
edges that are covered by the df relation.

Needs debugging. It looks like the call malloc -> bitcast ctrl-flow edge was
pruned, but not the others that still visible in the output on the synthetic
flowgraph test case.