Test if subgraphs are constant functions of the input
Walk back through the edges on the graph to decide if a value is derived entirely
from the function arguments and constants, without reference to memory. If so this
implies that we can propagate the model of the region produced back to the call-site
(if the region escapes the function).

Expanded the flowgraph test case to include static and dynamic arguments to the target
call.

Need to work out why %0 is spilled into a memory location, this may be a default choice
for function arguments to help the later register allocation pass but it is unnecessary
and blocks the analysis here without using aliasing info.