Previous problem was the optnone attribute
Clang emits this when -S is supplied to extract the source. This attribute is loaded
back in and stopped the pass-manager from doing anything useful.

Little shell to strip it out and the arguments to memory store is removed from the
function entry. Probably it is there to be conservative and one of the passes correctly
identifies tha the location is not aliased and the argument is not updated so it rewrites
the store/loads out of the chain.

The function body now looks correct - the loop has been unrolled and store propagation has
removed the memory accesses. My checker still does not identifiy them as static yet...