Extract the objects for vmlinux out of the make trace
Finds the line that calls the link-vmlinux.sh build script and extracts the arguments
passed (the list of objects and archives). Each archive is resolved by finding the
build step in the trace and decoding its arguments. This proceeds recursively until
the list of object files reaches a fixed point.

There are still some problems with the units.list produced - assembly generation
steps are not accounted for and there seem to be some units missing. The exraction and
rebuilding of .ll units needs to match the list generated.