Fixed bug in findobjs
Improved the funclist output by sorting the missing symbols, allows easier correlation
between modules that could not be loaded and external stubs. Used it to find a bug in
recursing through the kernel built-in archives.

Now the kernel import is very slow :) but the only stubs are assembly routines and llvm
library imports.