Put missing character classes into the regex matcher.
The callgrep binary now works correctly over the linux kernel. Two of the constructs
were missing from the regex matcher. This patch adds character classes and choice
will be put in later.

The performance is terrible. It takes two minutes to run callgrep over the kernel.
No proper profiling yet but most of the time is spent performing the link step from
the list of .ll inputs. About to experiment to see how much of this can be removed
by linking the .lls first...