Speed up tools on the linux test-case
Simulate the link step when the test-case is prepared instead of at beginning of each tool
run. The .ll files for the kernel are 2.3G while the linked .bc is "only" 464MB. This
takes a minute off of the tool start-up time although the rest of the module preparation
and call-graph building is still expensive.