Fixed a missing case in the regex matcher
Whena  node name has multiple choices, at the point that the match fails scan forwards in the
regex to check for more choices.

Now the linux kernel is working for the test query to find allocation site:
../../callgrep/build.sh && ../../callgrep/a.out '.*>_*[kzvm]*alloc|kmalloc_array.*' kernel.bc

The kmalloc_array function is marked as a static inline so some of the call-sites are
optimized away to the residual __kmalloc calls and some are preserved as kmalloc_array.xxxx
instances.